integrated new share functionality into core package
