+++ clowdr/__init__.py
@@ -4,5 +4,6 @@ from .task import TaskHandler
 from .controller import metadata, launcher, rerunner
 from .endpoint import AWS, remote
 from .driver import local, cloud, share
+from .share import consolidate, portal, customDash
 
-__all__ = ['task', 'controller', 'driver', 'endpoint']
+__all__ = ['task', 'controller', 'driver', 'endpoint', 'share']
