--- sharebox/README.md
@@ -1,5 +1,0 @@
-## Plotly Dash Experiments
-
-https://dash.plot.ly/
-
-
