--- sharebox/assets/banner.css
@@ -1,19 +1,0 @@
-body:after{
-  content: "alpha";
-  position: fixed;
-  width: 80px;
-  height: 25px;
-  background: #EE8E4A;
-  top: 7px;
-  right: -20px;
-  text-align: center;
-  font-size: 13px;
-  font-family: sans-serif;
-  text-transform: uppercase;
-  font-weight: bold;
-  color: #fff;
-  line-height: 27px;
-  -ms-transform:rotate(45deg);
-  -webkit-transform:rotate(45deg);
-  transform:rotate(45deg);
-}
