--- sharebox/gettingstarted2.py
@@ -1,80 +1,0 @@
-import dash
-import dash_core_components as dcc
-import dash_html_components as html
-from dash.dependencies import Input, Output
-
-app = dash.Dash()
-
-app.layout = html.Div(style={'backgroundColor': '#ffffff'}, children=[
-    html.H1(
-        children='Hello Dash',
-        style={
-            'textAlign': 'center',
-        }
-    ),
-    html.Div([
-        html.Label('Dropdown'),
-        dcc.Dropdown(
-            options=[
-                {'label': 'New York City', 'value': 'NYC'},
-                {'label': u'Montréal', 'value': 'MTL'},
-                {'label': 'San Francisco', 'value': 'SF'}
-            ],
-            value='MTL'
-        ),
-        html.Label('Multi-Select Dropdown'),
-        dcc.Dropdown(
-            options=[
-                {'label': 'New York City', 'value': 'NYC'},
-                {'label': u'Montréal', 'value': 'MTL'},
-                {'label': 'San Francisco', 'value': 'SF'}
-            ],
-            value=['NYC', 'SF'],
-            multi=True,
-            id="mydropd"
-        ),
-
-        html.Label('Radio Items'),
-        dcc.RadioItems(
-            options=[
-                {'label': 'New York City', 'value': 'NYC'},
-                {'label': u'Montréal', 'value': 'MTL'},
-                {'label': 'San Francisco', 'value': 'SF'}
-            ],
-            value='MTL'
-        ),
-
-        html.Label('Checkboxes'),
-        dcc.Checklist(
-            options=[
-                {'label': 'New York City', 'value': 'NYC'},
-                {'label': u'Montréal', 'value': 'MTL'},
-                {'label': 'San Francisco', 'value': 'SF'}
-            ],
-            values=['MTL', 'SF']
-        ),
-
-        html.Label('Text Input'),
-        dcc.Input(value='MTL', type='text', id='mytext'),
-
-        html.Label('Slider'),
-        dcc.Slider(
-            min=0,
-            max=9,
-            marks={i: 'Label {}'.format(i) if i == 1 else str(i) for i in range(1, 6)},
-            value=5,
-        )], style={"backgroundColor": "#ffffff", "columnCount": 2})
-])
-app.css.append_css({"external_url": "https://codepen.io/chriddyp/pen/bWLwgP.css"})
-
-
-@app.callback(
-    Output(component_id='mytext', component_property='value'),
-    [Input(component_id='mydropd', component_property='value')]
-)
-def update_output_div(input_value):
-    return ', '.join(input_value)
-
-
-if __name__ == '__main__':
-    app.run_server(debug=True)
